Changing double to in and raw value to decimal

+++ Domain/Models/PressureReading.cs
@@ -5,9 +5,9 @@ namespace Domain.Models
     public class PressureReading
     {
         public long Id { get; set; }
-        public int RawValue { get; set; }
-        public double PSI { get; set; }
-        public double BAR { get; set; }
+        public decimal RawValue { get; set; }
+        public decimal PSI { get; set; }
+        public decimal BAR { get; set; }
         public DateTimeOffset TimeStamp { get; set; }
         public string SensorName { get; set; }        
     }
